% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbPalette.R
\name{cbPalette}
\alias{cbPalette}
\title{Generate an actual colour palette from our simple list of colours}
\usage{
cbPalette()
}
\value{
a color palette
}
\description{
Generate an actual colour palette from our simple list of colours
}
\examples{
mypalette <- cbPalette()
}
