% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbPalette.R
\name{cvi_palettes}
\alias{cvi_palettes}
\title{Generate an actual colour palette from our simple list of colours}
\usage{
cvi_palettes(
  name,
  n,
  all_palettes = cbColors,
  type = c("discrete", "continuous")
)
}
\arguments{
\item{name}{Name of the color palette we want to use}

\item{n}{How many colors to use (default to full length of palette)}

\item{all_palettes}{the list of colour palettes we want to extract our choice from}

\item{type}{discrete or continuous palette}
}
\value{
a color palette
}
\description{
Generate an actual colour palette from our simple list of colours
}
\examples{
cbColors <- list(
  cvi_purples = c("#381532", "#4b1b42", "#5d2252", "#702963",
                "#833074", "#953784", "#a83e95"),
  cb7 = c('red', 'blue', 'grey39', 'lightsteelblue2', 'orange', 'black', 'pink3')
)
mypalette <- cvi_palettes("cb7", type = "discrete")
}
