% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbPalette.R
\name{scale_colour_cb}
\alias{scale_colour_cb}
\title{Function for ggplot to recognize palette (colour, for geom_point)}
\usage{
scale_colour_cb()
}
\value{
ggplot palette
}
\description{
Function for ggplot to recognize palette (colour, for geom_point)
}
\examples{
library(ggplot2)
months <- 24
person_names <- c("Adam", "Matt", "Sarah", "Ben", "Katie", "Cara", "Bob")
df <- expand.grid(month = 1:months, person = person_names)
initial_weights <- runif(length(person_names), min = 120, max = 180)
weights_map <- setNames(initial_weights, person_names)
df$weight <- weights_map[as.character(df$person)]
df$weight <- df$weight * (1 + runif(nrow(df), -0.1, 0.1))
df$person <- factor(df$person)
plt <- ggplot(df, mapping = aes(x = month, y = weight)) +
scale_y_continuous(limits = c(100, 230))
plt +
geom_point(aes(colour = person), size = 3) +
  geom_line(aes(color = person), linewidth = 1) +
  scale_colour_cb()
}
